requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
email-validator
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
python-dotenv
typing-extensions
